@@ -1,13 +1,17 @@
 import { Application } from "express";
-
-const hbs = require("hbs");
-const path = require("path");
+import { render } from "mustache";
+import fs from "fs";
+import path from "path";
 
 export const registerViews = (app: Application, rootDir: string) => {
-    app.set("view engine", "hbs");
-    app.set("views", path.join(rootDir, "views"));
-
-    hbs.registerHelper("json", (context: any) => {
-        return JSON.stringify(context);
+    app.engine("mustache", (filePath, options, callback) => {
+        try {
+            const mustacheTemplate = fs.readFileSync(filePath).toString();
+            return callback(null, render(mustacheTemplate, options));
+        } catch (err) {
+            return callback(err);
+        }
     });
+    app.set("view engine", "mustache");
+    app.set("views", path.join(rootDir, "views"));
 };
