@@ -1,7 +1,7 @@
 import { ConfigController } from "../controllers/configController";
+import { Router } from "express";
 
-const router = require("express").Router();
-
+const router = Router();
 router.get("/:appName", ConfigController.getConfig);
 
 export default router;
